Changed folder structure and distributed code to match the original modular intentions (no longer shove everyting in kernel, we've got handy file separation now)
