+++ srcCode/kernelModuleSrc/kernel/kernel.c
@@ -1,0 +1,48 @@
+// we're currently in protected real mode 32bit
+//full access to address space :3
+// apparently a stack pointer has been helpfully set by GRUB already. You can use stack - related C functions.
+//currently in a nakey environment - just cpu , no libc, no bios calls.
+
+
+#include <stddef.h>
+#include <stdint.h> // implements fixed-width integer types int8_t, uint16_t, int32_t, and uint64_t (among others). 
+ #include "terminal.h"  // "hey, my terminal functions are here!"
+#include "multiboot.h"  //later defined - multiboot protocol. needed for mem. map
+
+
+
+//multiboot mem map entry (from GRUB docs)
+typedef struct multiboot_mmap_entry {
+    uint32_t size;
+    uint64_t addr;
+    uint64_t len;
+    uint32_t type;
+} __attribute__((packed)) multiboot_mmap_entry_t;
+
+typedef struct multiboot_info {
+    uint32_t flags;
+    uint32_t mem_lower;
+    uint32_t mem_upper;
+    uint32_t boot_device;
+    uint32_t cmdline;
+    uint32_t mods_count;
+    uint32_t mods_addr;
+    uint32_t syms[4];
+    uint32_t mmap_length;
+    uint32_t mmap_addr;
+  
+} __attribute__((packed)) multiboot_info_t;
+
+
+
+//main 
+void kmain(void) {
+
+
+    terminal_initialize();
+    terminal_write("Hello, world from kernel!\n");
+    terminal_write("This is a new line!\n");
+    terminal_write("Nice, right?");
+    for (;;) {}
+}
+
